Factor out a few functions to super class UserView
