+++ ui/fmView.py
@@ -37,10 +37,16 @@ class FinancialManagerView(UserView):
             elif req_id in self.eventRequest_Control.get_id_of_event_request_for_user():
                 clear()
                 print(self.eventRequest_Control.get_reviews_from_event_request(id=req_id))
-                key=input("\nEnter financial review or [0] to go back: \n")
-                if not key=="0": self.eventRequest_Control.add_review(id=req_id, review=key)
+                
+                while(True):
+                    review = input("Please enter review: ")
+                    if(not self.verify_input()):
+                        continue
+                    self.eventRequest_Control.add_review(id=req_id, review=review)
+                    break
         return
 
+    # override
     def show_current_financial_requests(self): 
         self.financialRequest_Control=FinancialRequest_Control()
         while True:
@@ -53,7 +59,12 @@ class FinancialManagerView(UserView):
             elif req_id in self.financialRequest_Control.get_id_of_financial_request_for_user():
                 clear()
                 print(self.financialRequest_Control.get_str_financial_request_from_id(id=req_id))
-                key=input("\nEnter financial review or [0] to go back: \n")
-                if not key=="0": self.financialRequest_Control.add_review(id=req_id, review=key)
+                
+                while(True):
+                    review = input("Please enter review: ")
+                    if(not self.verify_input()):
+                        continue
+                    self.financialRequest_Control.add_review(id=req_id, review=review)
+                    break
         return
